Module arg_processing extracted from package
